+++ changelog.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - [PR44](https://github.com/scipopt/SCIPpp/pull/44), [Issue29](https://github.com/scipopt/SCIPpp/issues/29) Support
   of ObjSCIP without accessing the raw SCIP object:
   - message handlers derived from `scip::ObjMessagehdlr` via `Model::setMessagehdlr`,
+  - problem data derived from `scip::ObjProbData` via a new constructor of `Model`,
   - Benders' decompositions derived from `scip::ObjBenders` via `Model::includeBenders`,
   - Benders' decomposition cuts derived from `scip::ObjBenderscut` via `Model::includeBenderscut`,
   - branching rules derived from `scip::ObjBranchrule` via `Model::includeBranchrule`,
@@ -18,10 +19,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - event handlers derived from `scip::ObjEventhdlr` via `Model::includeEventhdlr`,
   - file readers derived from `scip::ObjReader` via `Model::includeReader`,
   - IIS finders derived from `scip::ObjIISfinder` via `Model::includeIISfinder`,
-  - primal heuristics derived from `scip::ObjHeur` via `Model::includeHeur`,
   - node selectors derived from `scip::ObjNodesel` via `Model::includeNodesel`,
   - presolvers derived from `scip::ObjPresol` via `Model::includePresol`,
   - pricers derived from `scip::ObjPricer` via `Model::includePricer`,
+  - primal heuristics derived from `scip::ObjHeur` via `Model::includeHeur`,
   - propagators derived from `scip::ObjProp` via `Model::includeProp`,
   - relaxators derived from `scip::ObjRelax` via `Model::includeRelax`,
   - separators derived from `scip::ObjSepa` via `Model::includeSepa`, and